requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
streamlit==1.38.0
python-dotenv==1.0.1
google-generativeai==0.8.5
